--- batch/src/main/java/store/_0982/batch/application/settlement/SellerSettlementProcessor.java
@@ -1,32 +1,0 @@
-package store._0982.batch.application.settlement;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import store._0982.batch.infrastructure.client.product.ProductClient;
-import store._0982.batch.infrastructure.client.product.dto.GroupPurchaseInternalInfo;
-
-import java.util.List;
-import java.util.UUID;
-
-@RequiredArgsConstructor
-@Service
-public class SellerSettlementProcessor {
-
-    private final SellerBalanceCommandService sellerBalanceCommandService;
-    private final ProductClient productClient;
-
-    @Transactional
-    public void processSellerSettlement(UUID sellerId, List<GroupPurchaseInternalInfo> groupPurchases) {
-
-        Long totalAmount = groupPurchases.stream()
-                .mapToLong(gp -> gp.totalAmount() != null ? gp.totalAmount() : 0L)
-                .sum();
-
-        sellerBalanceCommandService.increaseBalance(sellerId, totalAmount);
-
-        for (GroupPurchaseInternalInfo gp : groupPurchases) {
-            productClient.markAsSettled(gp.groupPurchaseId());
-        }
-    }
-}
